Refactored the restoration code into seperate modules to allow easier intergration of other rstoration requirements
