+++ src/aims_ui/templates/aims_ui_pages/a_single_matches/singlesearch.html
@@ -11,6 +11,7 @@
 {% from 'custom_download_results_from_local_storage/_macro.njk' import customDownloadResultsFromLocalStorage %}
 {% from 'custom_we_matched_message/_macro.njk' import weMatchedAddressPreamble %}
 {% from 'custom_additional_request_info/_macro.njk' import customAdditionalRequestInfo %}
+{% from 'custom_run_script_with_page_name/_macro.njk' import runScriptWithPageName %}
 
 {# In order to specify wider than usual components, page must be in "pageContent" #}
 {% block pageContent %}
@@ -31,6 +32,7 @@
 		   "endpoints": endpoints,
 		   "page_name": page_name,
 		   "remove_search_button":False,
+			 "component_wrappers": True,
        "results_page": results_page, }) }}
 <br>
 </div>
@@ -88,7 +90,12 @@
 {# Specific page helpers here: (commented out if none required) #}
 {# <script type=module src="static/js/a_single_matches/singlesearch.mjs"></script> #}
 
-<script type=module src="static/js/e_all_pages/save_inputs.mjs"></script>
+<!-- Save and load inputs with the new mechanism -->
+{{ runScriptWithPageName({
+		"page_name": page_name,
+		"script_location": '/static/js/e_all_pages/save_and_restore_inputs.mjs',
+}) }}
+
 <script type=module src="static/js/e_all_pages/custom_column_widths.mjs"></script>
 
 <script type=module src="static/js/e_all_pages/all_pages_last.mjs"></script>
