+++ src/aims_ui/templates/aims_ui_pages/d_misc_functionality/settings_snippets/_settings_persistance.html
@@ -39,7 +39,7 @@
 
         <td class="ons-table__cell">
         {{ onsCheckbox({
-            "id": "singlesearch-classification-checkbox",
+            "id": "singlesearch-classificationfilter-checkbox",
             "value": "true",
             "hideLabel": true,
             "label": { "text": "" } }) }} </td>
